Buttons only appear when user is admin

+++ app/src/main/java/hbv601g/hugb2_team2/ui/activities/main/fragments/establishment_list_fragment/EstablishmentAdapter.kt
@@ -62,14 +62,14 @@ class EstablishmentAdapter(
         }
 
         // Delete button visibility and action
-        /*if (sessionManager.isLoggedIn() || sessionManager.isAdmin()) {
+        if (sessionManager.isLoggedIn() || sessionManager.isAdmin()) {
             holder.buttonDelete.visibility = View.VISIBLE
-            holder.buttonDelete.setOnClickListener {
+           /* holder.buttonDelete.setOnClickListener {
                 onDeleteClicked(establishment)
-            }
+            }*/
         } else {
             holder.buttonDelete.visibility = View.GONE
-        }*/
+        }
 
     }
